IOS-6890 Additional files
