+++ TangemStaking/Models/RewardClaimingType.swift
@@ -1,0 +1,14 @@
+//
+//  RewardClaimingType.swift
+//  TangemStaking
+//
+//  Created by Sergey Balashov on 27.05.2024.
+//  Copyright © 2024 Tangem AG. All rights reserved.
+//
+
+import Foundation
+
+public enum RewardClaimingType: String, Hashable {
+    case auto
+    case manual
+}
