+++ TangemStaking/Models/RewardType.swift
@@ -1,0 +1,15 @@
+//
+//  RewardType.swift
+//  TangemStaking
+//
+//  Created by Sergey Balashov on 27.05.2024.
+//  Copyright © 2024 Tangem AG. All rights reserved.
+//
+
+import Foundation
+
+public enum RewardType: String, Hashable {
+    case apr
+    case apy
+    case variable
+}
