+++ TangemStaking/Models/YieldInfo.swift
@@ -7,7 +7,6 @@
 //
 
 import Foundation
-import BlockchainSdk
 
 public struct YieldInfo {
     let item: StakingTokenItem
@@ -19,39 +18,4 @@ public struct YieldInfo {
     let rewardClaimingType: RewardClaimingType
     let warmupPeriod: Period
     let rewardScheduleType: RewardScheduleType
-}
-
-public enum Period {
-    case days(_ days: Int)
-
-    // TODO: To UI layer
-    func formatted(formatter: DateComponentsFormatter) -> String {
-        switch self {
-        case .days(let days):
-            formatter.unitsStyle = .short
-            formatter.allowedUnits = [.day]
-            return formatter.string(from: DateComponents(day: days)) ?? days.formatted()
-        }
-    }
-}
-
-public enum RewardClaimingType: String, Hashable {
-    case auto
-    case manual
-}
-
-public enum RewardScheduleType: String, Hashable {
-    case block
-    case hour
-    case day
-    case week
-    case month
-    case era
-    case epoch
-}
-
-public enum RewardType: String, Hashable {
-    case apr
-    case apy
-    case variable
 }
